Правки проблем из списка FindBugs #2

+++ CodingDojo/games/engine/src/main/java/com/codenjoy/dojo/utils/TestUtils.java
@@ -26,10 +26,10 @@ package com.codenjoy.dojo.utils;
 import com.codenjoy.dojo.services.EventListener;
 import com.codenjoy.dojo.services.Game;
 import com.codenjoy.dojo.services.GameType;
-import com.codenjoy.dojo.services.printer.PrinterFactory;
 import com.codenjoy.dojo.services.multiplayer.GameField;
 import com.codenjoy.dojo.services.multiplayer.GamePlayer;
 import com.codenjoy.dojo.services.multiplayer.Single;
+import com.codenjoy.dojo.services.printer.PrinterFactory;
 
 public class TestUtils {
 
@@ -41,7 +41,7 @@ public class TestUtils {
     }
 
     public static String injectNN(String expected) {
-        int size = (int) Math.sqrt(expected.length()/COUNT_NUMBERS)*COUNT_NUMBERS;
+        int size = (int) Math.sqrt(expected.length()/(double)COUNT_NUMBERS)*COUNT_NUMBERS;
         return inject(expected, size, "\n");
     }
